seperate category helper funcs from methods and update imports
